Uncommited Changes

+++ src/main/java/me/suff/mc/regen/common/entities/TimelordEntity.java
@@ -108,7 +108,7 @@ public class TimelordEntity extends VillagerEntity implements IRangedAttackMob {
     public static AttributeModifierMap.MutableAttribute createAttributes() {
         return MonsterEntity.createMonsterAttributes().
                 add(Attributes.FOLLOW_RANGE, 35D).
-                add(Attributes.MOVEMENT_SPEED, (double) 0.23F).
+                add(Attributes.MOVEMENT_SPEED, 0.23F).
                 add(Attributes.ATTACK_DAMAGE, 3F).
                 add(Attributes.MAX_HEALTH, 20D).
                 add(Attributes.ARMOR, 2.0D);
